fix clippy

+++ components/raftstore/src/store/worker/raftlog_fetch.rs
@@ -80,7 +80,7 @@ where
                 tried_cnt,
                 term,
             } => {
-                let mut ents = Vec::with_capacity(high - low);
+                let mut ents = Vec::with_capacity((high - low) as usize);
                 let res = self.raft_engine.fetch_entries_to(
                     region_id,
                     low,
